chore(ClientList): Refactor where the request is
BREAKING CHANGE: The ClientListRequest moved into its own namespace.
